requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
Brotli==1.0.9
bs4==0.0.1
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
configparser==5.3.0
contourpy==1.0.6
cryptography==39.0.0
cycler==0.11.0
dash==2.7.0
dash-core-components==1.0.0
dash-html-components==1.0.0
dash-renderer==1.0.0
dash-table==4.0.0
dataframe-image==0.1.3
DateTime==4.9
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.8.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.4
fastjsonschema==2.16.2
filelock==3.8.0
Flask==2.2.2
fonttools==4.38.0
frozendict==2.3.4
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.30
gunicorn==20.1.0
html5lib==1.1
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.1
jsonschema==3.2.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.6.2
missingno==0.5.1
mistune==2.0.4
multidict==6.0.3
multimethod==1.9.1
multitasking==0.0.11
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
nltk==3.8
numpy==1.23.5
oauthlib==3.2.2
packaging==21.3
pandas==1.5.2
pandas-profiling==3.2.0
pandocfilters==1.5.0
paramiko==2.12.0
pathlib==1.0.1
phik==0.12.3
Pillow==9.3.0
pipreqs==0.4.11
platformdirs==2.5.4
plotly==5.11.0
pluggy==1.0.0
protobuf==3.20.3
psycopg2-binary==2.9.5
pyarrow==10.0.1
pycparser==2.21
pydantic==1.10.4
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==305
PyYAML==5.4.1
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.0.0
scipy==1.9.3
seaborn==0.12.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
streamlit==1.11.1
streamlit-ace==0.1.1
streamlit-discourse==0.1.3
streamlit-disqus==0.1.3
streamlit-elements==0.1.0
streamlit-on-Hover-tabs==1.0.1
streamlit-pandas-profiling==0.1.3
streamlit-player==0.1.5
streamlit-quill==0.0.3
tangled-up-in-unicode==0.2.0
tenacity==8.1.0
termcolor==2.1.1
textblob==0.17.1
texttable==1.6.7
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
tweepy==4.4.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
virtualenv==20.17.0
visions==0.7.4
waitress==2.1.2
watchdog==2.2.1
webencodings==0.5.1
websocket-client==0.59.0
Werkzeug==2.2.2
yarg==0.1.9
yarl==1.8.2
yfinance==0.2.3
zipp==3.11.0
zope.interface==5.5.2
